Switched add_numbers tool to int also for Gradio for consistency (#23)

+++ python_examples/ai/mcp/gradio/gradio_mcp_server/server.py
@@ -3,7 +3,7 @@ import gradio as gr
 from gradio.mcp import GradioMCPServer
 
 
-def add_numbers(a: float, b: float) -> float:
+def add_numbers(a: int, b: int) -> int:
     """Add two numbers"""
     return a + b
